- Added "category" property to Car, - Updated filterCars endpoint to filter cars by category, price range and location radius as well (the latest needs to be tested better)

+++ src/main/kotlin/com/rentmycar/entities/Location.kt
@@ -9,8 +9,8 @@ import org.jetbrains.exposed.dao.*
 
 object Locations : IntIdTable() {
     val car = reference("car_id", Cars)
-    val longitude = float("longitude")
-    val latitude = float("latitude")
+    val longitude = double("longitude")
+    val latitude = double("latitude")
 }
 
 class Location(id: EntityID<Int>) : IntEntity(id) {
